+++ src/main/javascript/project/index.js
@@ -3,6 +3,7 @@
 const path = require('path');
 const manifestSchema = require('../../resources/manifest.schema.json');
 const DeskproProject = require('./DeskproProject');
+const DeskproApiClient = require('./DeskproApiClient');
 
 const source = path.resolve(__dirname, "../../resources/scaffold");
 
@@ -13,5 +14,12 @@ exports = module.exports = {
      */
     newInstance: function () {
         return new DeskproProject(manifestSchema, source);
+    },
+    /**
+     * @param {String} dpInstanceUrl
+     * @returns {DeskproApiClient}
+     */
+    newApiClientInstance: function (dpInstanceUrl) {
+        return DeskproApiClient.createInstance(dpInstanceUrl);
     }
 };