+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Upcoming
+
+* Add package and deploy commands
+
 ## v0.1.0 - 2017-04-24
 
 * Initial public release